refactor(rag): replace the index status webhook with a route
The RAG indexer reported the indexation status of a file by posting on a
per-instance @webhook trigger, which pushed a job to the rag-index-status
worker. The callback is now handled synchronously by a route, so the indexer
gets a real status code instead of a 204 sent before its payload was even read.

SetRAGStatus is unchanged.
